new_skraggle/requirements.txt
alembic==1.7.5
amqp==5.1.1
APScheduler==3.8.1
attrs==21.4.0
Authlib==0.15.5
Babel==2.9.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
blinker==1.4
cachelib==0.5.0
cachetools==4.2.4
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crypto==1.4.1
cryptography==35.0.0
dat==1.0.0
dataclasses-json==0.5.6
Deprecated==1.2.13
dnspython==2.2.1
e==1.4.5
email-validator==1.2.1
et-xmlfile==1.1.0
facebook-sdk==3.1.0
first==2.0.2
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-BabelEx==0.9.4
Flask-Caching==1.10.1
Flask-Classy==0.6.10
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-Pydantic==0.9.0
Flask-Security==3.0.0
Flask-Session==0.4.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google==3.0.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
kombu==5.2.4
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
Naked==0.1.31
nulltype==2.3.1
numpy==1.21.5
oauth2==1.9.0.post1
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
parameters==0.2.1
passlib==1.7.4
paypalrestsdk==1.13.1
plaid-python==8.6.0
pluggy==1.0.0
prompt-toolkit==3.0.29
protobuf==3.19.1
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.12.0
pydantic==1.8.2
PyJWT==2.0.0
pymongo==4.0.1
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.6
pytest==7.1.2
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-facebook-api==0.12.4
python-http-client==3.3.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
resp==0.1.2
rq==1.10.1
rsa==4.7.2
salesforce==0.0.1
sendgrid==6.9.3
shellescape==3.8.1
shortuuid==1.0.8
simple-salesforce==1.11.4
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
speaklater==1.3
SQLAlchemy==1.4.29
sqlalchemy-mixins==1.5.3
starkbank-ecdsa==2.0.3
stripe==2.61.0
tomli==2.0.1
tweepy==4.4.0
typing-inspect==0.7.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
username==2.1.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
zipp==3.6.0
